+++ custom_components/houseplan/frontend/houseplan-card.js
@@ -1,1 +1,1 @@
-globalThis.__HOUSEPLAN_BUILD_FINGERPRINT__="13fcebb68eb62b100bccb7b77eb09b8fce2f41e6c77bcbe8870e6dc4d97eedbf";export{d6 as HouseplanCard}from"./houseplan-assets/houseplan-card-DqEPY1YF.js";
+globalThis.__HOUSEPLAN_BUILD_FINGERPRINT__="0c61742c98d55aec077478fcee7b8acf52e2ac1b70e2ed23532483b845ca884f";export{d6 as HouseplanCard}from"./houseplan-assets/houseplan-card-BU843YpJ.js";